refactor(number-input): Separation of `NumberInput` and `Input`
