+++ src/form/input/input.test.tsx
@@ -70,95 +70,3 @@ describe("<Input />", () => {
     expect(input).toBeDisabled();
   });
 });
-
-describe('<Input type={"number"} />', () => {
-  const numberInputProps: InputProps = {
-    testid: "number-input",
-    name: "number-input",
-    type: "number",
-    localizationOptions: {
-      shouldFormatToLocaleString: true,
-      locale: "en-EN",
-      maximumFractionDigits: 2
-    },
-    onChange: jest.fn()
-  };
-
-  it("should render correctly", () => {
-    render(<Input {...numberInputProps} />);
-  });
-
-  it("should pass a11y test", async () => {
-    const {container} = render(<Input {...numberInputProps} />);
-
-    await testA11y(container, {rules: {label: {enabled: false}}});
-  });
-
-  it("should format value to with 'en-EN' locale", () => {
-    render(<Input {...numberInputProps} value={"1234567.89"} />);
-
-    const input = screen.getByRole("textbox");
-
-    expect(input).toHaveValue("1,234,567.89");
-  });
-
-  it("should remove leading zeros", () => {
-    render(<Input {...numberInputProps} value={"0001234.50"} />);
-
-    const input = screen.getByRole("textbox");
-
-    expect(input).toHaveValue("1,234.50");
-  });
-
-  it("should parse to scientific notation", () => {
-    render(<Input {...numberInputProps} />);
-
-    const input = screen.getByRole("textbox");
-
-    userEvent.type(input, "1,234,567.89");
-
-    expect(input).toHaveValue("1234567.89");
-  });
-
-  it("should have at most 2 decimal places", () => {
-    render(<Input {...numberInputProps} />);
-
-    const input = screen.getByRole("textbox");
-
-    userEvent.type(input, "100.55555");
-
-    expect(input).toHaveValue("100.55");
-  });
-
-  it("should not allow enter letter", () => {
-    render(<Input {...numberInputProps} />);
-
-    const input = screen.getByRole("textbox");
-
-    userEvent.type(input, "ABC");
-
-    expect(input).toHaveValue(undefined);
-  });
-
-  it("should not allow enter empty string", () => {
-    render(<Input {...numberInputProps} />);
-
-    const input = screen.getByRole("textbox");
-
-    userEvent.type(input, " ");
-
-    expect(input).toHaveValue(undefined);
-  });
-
-  it("should not allow negative zero without decimal part", () => {
-    render(
-      <Input {...numberInputProps} localizationOptions={{maximumFractionDigits: 0}} />
-    );
-
-    const input = screen.getByRole("textbox");
-
-    userEvent.type(input, "-0");
-
-    expect(input).toHaveValue("0");
-  });
-});
