+++ src/form/input/number/util/numberInputTypes.ts
@@ -1,0 +1,11 @@
+import {InputProps} from "../../util/inputTypes";
+
+export type NumberInputFormatProps = {
+  locale?: string;
+  shouldFormatToLocaleString?: boolean;
+};
+
+export type NumberInputProps = Omit<InputProps, "type"> & {
+  formatProps?: NumberInputFormatProps;
+  maximumFractionDigits?: number;
+};
