+++ stories/4-NumberInput.stories.tsx
@@ -4,21 +4,20 @@ import {storiesOf} from "@storybook/react";
 import StateProvider from "./utils/StateProvider";
 
 import FormField from "../src/form/field/FormField";
-import Input from "../src/form/input/Input";
+import NumberInput from "../src/form/input/number/NumberInput";
 
 storiesOf("Number Input", module)
   .add("Number Input", () => (
     <Fragment>
-      <StateProvider initialState={""}>
+      <StateProvider initialState={{value: ""}}>
         {(state, setState) => (
           <Fragment>
             <FormField label={"Price – maximumFractionDigits={2}"}>
-              <Input
-                localizationOptions={{maximumFractionDigits: 2}}
+              <NumberInput
+                maximumFractionDigits={2}
                 name={"price"}
-                type={"number"}
-                onChange={(e) => setState(e.currentTarget.value)}
-                value={state}
+                onChange={(e) => setState({value: e.currentTarget.value})}
+                value={state.value}
                 placeholder={"$ 10"}
               />
             </FormField>
@@ -28,7 +27,7 @@ storiesOf("Number Input", module)
         )}
       </StateProvider>
 
-      <StateProvider initialState={""}>
+      <StateProvider initialState={{value: ""}}>
         {(state, setState) => (
           <Fragment>
             <br />
@@ -36,12 +35,11 @@ storiesOf("Number Input", module)
             <FormField
               label={"Price - Has Error"}
               errorMessages={["Please enter a valid price"]}>
-              <Input
-                localizationOptions={{maximumFractionDigits: 2}}
+              <NumberInput
+                maximumFractionDigits={2}
                 name={"price"}
-                type={"number"}
                 onChange={(e) => setState(e.currentTarget.value)}
-                value={state}
+                value={state.value}
                 placeholder={"$ 10"}
                 hasError={true}
               />
@@ -55,16 +53,15 @@ storiesOf("Number Input", module)
   ))
   .add("Number Input with maximumFractionDigit", () => (
     <Fragment>
-      <StateProvider initialState={""}>
+      <StateProvider initialState={{value: ""}}>
         {(state, setState) => (
           <Fragment>
             <FormField label={"Price (BTC) – maximumFractionDigits={8}"}>
-              <Input
-                localizationOptions={{maximumFractionDigits: 8}}
+              <NumberInput
+                maximumFractionDigits={8}
                 name={"price"}
-                type={"number"}
                 onChange={(e) => setState(e.currentTarget.value)}
-                value={state}
+                value={state.value}
                 placeholder={"Min. ₿ 0.00000001"}
               />
             </FormField>
@@ -76,18 +73,17 @@ storiesOf("Number Input", module)
 
       <br />
 
-      <StateProvider initialState={""}>
+      <StateProvider initialState={{value: ""}}>
         {(state, setState) => (
           <Fragment>
             <FormField
               label={
                 "ID or Passport Number – maxFractionDigits={0} - Can have leading zeros"
               }>
-              <Input
+              <NumberInput
                 name={"id-number"}
-                type={"number"}
                 onChange={(e) => setState(e.currentTarget.value)}
-                value={state}
+                value={state.value}
               />
             </FormField>
 
@@ -99,23 +95,22 @@ storiesOf("Number Input", module)
   ))
   .add("Number Input with shouldFormatToLocaleString", () => (
     <Fragment>
-      <StateProvider initialState={""}>
+      <StateProvider initialState={{value: ""}}>
         {(state, setState) => (
           <Fragment>
             <FormField
               label={
                 "Wallet ($) – maximumFractionDigits={2} - shouldFormatToLocaleString={true}"
               }>
-              <Input
+              <NumberInput
                 name={"wallet"}
-                localizationOptions={{
-                  maximumFractionDigits: 2,
+                formatProps={{
                   shouldFormatToLocaleString: true
                 }}
-                type={"number"}
+                maximumFractionDigits={2}
                 placeholder={"$ 1,000"}
                 onChange={(e) => setState(e.currentTarget.value)}
-                value={state}
+                value={state.value}
               />
             </FormField>
 
@@ -126,20 +121,19 @@ storiesOf("Number Input", module)
 
       <br />
 
-      <StateProvider initialState={""}>
+      <StateProvider initialState={{value: ""}}>
         {(state, setState) => (
           <Fragment>
             <FormField
               label={
                 "World Population – maximumFractionDigits={0} - shouldFormatToLocaleString={true}"
               }>
-              <Input
+              <NumberInput
                 name={"world-population"}
-                type={"number"}
                 placeholder={"7,794,798,739"}
-                localizationOptions={{shouldFormatToLocaleString: true}}
+                formatProps={{shouldFormatToLocaleString: true}}
                 onChange={(e) => setState(e.currentTarget.value)}
-                value={state}
+                value={state.value}
               />
             </FormField>
 
@@ -151,23 +145,22 @@ storiesOf("Number Input", module)
   ))
   .add("Number Input with locale", () => (
     <Fragment>
-      <StateProvider initialState={""}>
+      <StateProvider initialState={{value: ""}}>
         {(state, setState) => (
           <Fragment>
             <FormField
               label={
                 'Budget (Locale) – maximumFractionDigits={4} - shouldFormatToLocaleString={true} - locale={"zh-Hans-CN-u-nu-hanidec"}'
               }>
-              <Input
+              <NumberInput
                 name={"world-population"}
-                type={"number"}
-                localizationOptions={{
-                  maximumFractionDigits: 4,
+                maximumFractionDigits={4}
+                formatProps={{
                   shouldFormatToLocaleString: true,
                   locale: "zh-Hans-CN-u-nu-hanidec"
                 }}
                 onChange={(e) => setState(e.currentTarget.value)}
-                value={state}
+                value={state.value}
               />
             </FormField>
 
@@ -178,23 +171,22 @@ storiesOf("Number Input", module)
 
       <br />
 
-      <StateProvider initialState={""}>
+      <StateProvider initialState={{value: ""}}>
         {(state, setState) => (
           <Fragment>
             <FormField
               label={
                 'Budget (Locale) – maximumFractionDigits={4} - shouldFormatToLocaleString={true} - locale={"tr"}'
               }>
-              <Input
+              <NumberInput
                 name={"world-population"}
-                type={"number"}
-                localizationOptions={{
-                  maximumFractionDigits: 4,
+                maximumFractionDigits={4}
+                formatProps={{
                   shouldFormatToLocaleString: true,
                   locale: "tr"
                 }}
                 onChange={(e) => setState(e.currentTarget.value)}
-                value={state}
+                value={state.value}
               />
             </FormField>
